backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
httpx==0.28.1
redis[hiredis]==5.2.1
cryptography==44.0.0
python-multipart==0.0.19
email-validator==2.2.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
